feat: api to avaliation the books

+++ src/main/java/com/curvelo/repository/BookRepository.java
@@ -1,9 +1,9 @@
 package com.curvelo.repository;
 
-import com.curvelo.model.Book;
+import com.curvelo.domain.model.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer> {
 
   boolean existsByIsbn(String isbn);
